:sparkles: update quantity of products in cart

+++ views/cart.php
@@ -22,6 +22,13 @@ if (isset($_GET['action'])) {
         case 'delete_all':
             unset($_SESSION['cart']);
             break;
+        case 'update':
+            if ($_GET['type'] == 'asc')
+                $_SESSION['cart'][$id]++;
+            else if ($_GET['type'] == 'dec')
+                if ($_SESSION['cart'][$id] > 1) $_SESSION['cart'][$id]--;
+            header("location: ?option=cart");
+            break;
     }
 }
 ?>
@@ -50,9 +57,13 @@ if (isset($_GET['action'])) {
                 ?>
                     <tr>
                         <td width="20%"><img width="100%" src="./images/<?= $item['image'] ?>"></td>
-                        <td><?= $item['name'] ?><br><input type="button" value="Delete" onclick="location='?option=cart&action=delete&id=<?=$item['id']?>';"></td>
+                        <td><?= $item['name'] ?><br><input type="button" value="Delete" onclick="location='?option=cart&action=delete&id=<?= $item['id'] ?>';"></td>
                         <td><?= number_format($item['price'], 0, ',', '.') ?></td>
-                        <td><?= $_SESSION['cart'][$item['id']] ?></td>
+                        <td>
+                            <input type="button" value="-" onclick="location='?option=cart&action=update&type=dec&id=<?= $item['id'] ?>';">
+                            <?= $_SESSION['cart'][$item['id']] ?>
+                            <input type="button" value="+" onclick="location='?option=cart&action=update&type=asc&id=<?= $item['id'] ?>';">
+                        </td>
                         <td><?= number_format($subTotal = $item['price'] * $_SESSION['cart'][$item['id']], 0, ',', '.') ?></td>
                         <?php $total += $subTotal; ?>
                     </tr>
@@ -63,8 +74,8 @@ if (isset($_GET['action'])) {
                     <td colspan="5">
                         <section>
                             Total: <?= number_format($total, 0, ',', '.') ?> VND
-                            <input type="button" value="Delete All" onclick="location='?option=cart&action=delete_all'">
-                    </section> 
+                            <input type="button" value="Delete All" onclick="if(confirm('Are you want to delete all cart?')) location='?option=cart&action=delete_all'">
+                        </section>
                     </td>
                 </tr>
             </tbody>
